Release 1.03.09
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@538 98333e67-4a24-44d7-a75c-e53540dd3050

+++ branches/release_1_03/xmlrpc-c/src/xmlrpc_server_abyss.c
@@ -198,9 +198,9 @@ getBody(xmlrpc_env *        const envP,
 
     body = xmlrpc_mem_block_new(envP, 0);
     if (!envP->fault_occurred) {
-        unsigned int bytesRead;
+        size_t bytesRead;
         char * chunkPtr;
-        int chunkLen;
+        size_t chunkLen;
 
         bytesRead = 0;
